+++ lib/swiss_knife.dart
@@ -7,4 +7,5 @@ export 'src/collections.dart';
 export 'src/utils.dart';
 export 'src/data.dart';
 export 'src/resource.dart';
+export 'src/json.dart';